moviehub/requirements.txt
agate==1.11.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
asgiref==3.8.1
Babel==2.15.0
csvkit==2.0.0
dbfread==2.0.7
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==25.8.0
isodate==0.6.1
leather==0.4.0
numpy==1.26.4
olefile==0.47
openpyxl==3.1.3
pandas==2.2.2
parsedatetime==2.6
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
xlrd==2.0.1
